Gatepass/requirements.txt
Django==4.2.7
asgiref==3.7.2
sqlparse==0.4.4
tzdata==2023.3
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==1.2.0
djangorestframework==3.15.0
django-cors-headers==4.0.0
psycopg[binary]==3.3.2
openpyxl==3.1.2
